requirements.txt
art
opencv-python
numpy
pyautogui
pywin32
pyserial
dxcam
colorama
